+++ src/ascat/eumetsat/level2.py
@@ -17,6 +17,7 @@ from ascat.read_native.eps_native import AscatL2EpsFileGeneric
 from ascat.utils import get_toi_subset, get_roi_subset
 from ascat.utils import get_file_format
 from ascat.file_handling import ChronFiles
+from ascat.utils import Spacecraft
 
 
 class AscatL2File:
@@ -62,8 +63,8 @@ class AscatL2BufrFileList(ChronFiles):
         """
         Initialize.
         """
-        sat_lut = {"a": 2, "b": 1, "c": 3}
-        self.sat = sat_lut[sat]
+        sat_lut = {"A": 2, "B": 1, "C": 3}
+        self.sat = sat_lut[Spacecraft(sat).sat_name]
         self.product = product
 
         if filename_template is None:
@@ -126,7 +127,7 @@ class AscatL2NcFileList(ChronFiles):
             Filename template (default:
             "M0{sat}-ASCA-ASC{product}1B0200-NA-9.1-{date}.000000000Z-{placeholder1}-{placeholder2}.bfr")
         """
-        self.sat = sat
+        self.sat = Spacecraft(sat).sat_name
 
         lut = {"smr": "125", "smo": "250"}
         self.product = lut[product]
@@ -188,8 +189,8 @@ class AscatL2EpsFileList(ChronFiles):
             Filename template (default:
                 "ASCA_{product}_02_M0{sat}_{date}Z_*_*_*_*.nat")
         """
-        sat_lut = {"a": 2, "b": 1, "c": 3, "?": "?"}
-        self.sat = sat_lut[sat]
+        sat_lut = {"A": 2, "B": 1, "C": 3}
+        self.sat = "?" if sat == "?" else sat_lut[Spacecraft(sat).sat_name]
         self.product = product
 
         if filename_template is None: